llvm-flow-backend/requirements.txt
asgiref==3.5.2
charset-normalizer==2.1.0
DateTime==4.5
Django==3.2.5
django == 3.2.5
django-core==1.4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
environ==1.0
gunicorn==20.1.0 
idna==3.3
PyJWT==1.7.1
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
zope.interface==5.4.0
